+++ src/main/java/in/koreatech/koin/admin/student/service/AdminStudentService.java
@@ -11,7 +11,7 @@ import in.koreatech.koin.admin.student.dto.AdminStudentResponse;
 import in.koreatech.koin.admin.student.dto.AdminStudentUpdateRequest;
 import in.koreatech.koin.admin.student.dto.AdminStudentUpdateResponse;
 import in.koreatech.koin.admin.student.dto.AdminStudentsResponse;
-import in.koreatech.koin.admin.user.dto.StudentsCondition;
+import in.koreatech.koin.admin.student.dto.StudentsCondition;
 import in.koreatech.koin.admin.user.repository.AdminUserRepository;
 import in.koreatech.koin.domain.student.exception.StudentDepartmentNotValidException;
 import in.koreatech.koin.domain.student.model.Student;
